Resolve conflict and lint warning
1. updated the branch and resolved conflict.
2. resolve lint warning.
